Star Web App 2nd Draft
Updated this with some interactivity, animation and ux tweaks. Replaced links in nav with the ones on their current site.  I created the blog/news pages in bare form to be connected to the db. Login page is included and accessable from admin button. I also created all the blank pages needed to make it easier to finish. Everything is linked in the navbar though I have the buttons disabled right now. There is still plenty of things to fix and tweak, such as the footer for example and the remaining pages. This front end may have to be changed also if they do not agree with it, but now we have some barebones markup/css code to go off of.