Added to success stories, demonstrated hiding elements on main page
Learn more button on first page hides save changes button and disables editing, to show what admin functionality could be like. This is only to demo how we can implement that and not meant to be permanent. I still want to work on retrieving success stories and maybe fixing errors here and there.